refactor: 하나의 CommentController를 PostCommentController와 ProductCommentController로 분리
