+++ src/main/java/com/enjoytrip/comment/mapper/PostCommentMapper.java
@@ -4,6 +4,8 @@ import com.enjoytrip.comment.dto.PostCommentDto;
 import com.enjoytrip.comment.entity.PostComment;
 import com.enjoytrip.member.entity.Member;
 import com.enjoytrip.member.service.MemberService;
+import com.enjoytrip.post.entity.Post;
+import com.enjoytrip.post.service.PostService;
 import lombok.RequiredArgsConstructor;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -17,6 +19,7 @@ import java.util.List;
 public abstract class PostCommentMapper {
 
     private final MemberService memberService;
+    private final PostService postService;
 
     @Mapping(source = "id", target = "commentId")
     public abstract PostCommentDto.Get postCommentToGetRequest(PostComment postComment);
@@ -24,7 +27,8 @@ public abstract class PostCommentMapper {
     public abstract List<PostCommentDto.Get> postCommentToGetRequest(List<PostComment> postComment);
 
     @Mapping(target = "writer", source = "postRequest.writerId", qualifiedByName = "mapMember")
-    public abstract PostComment postRequestToPostComment(PostCommentDto.Post postRequest);
+    @Mapping(target = "post", source = "postId", qualifiedByName = "mapPost")
+    public abstract PostComment postRequestToPostComment(PostCommentDto.Post postRequest, Long postId);
 
     public abstract void patchRequestToPostComment(PostCommentDto.Patch patchRequest, @MappingTarget PostComment postComment);
 
@@ -32,6 +36,11 @@ public abstract class PostCommentMapper {
     @Named("mapMember")
     protected Member mapMember(Long memberId) {
         return memberService.findOneMember(memberId);
+    }
+
+    @Named("mapPost")
+    protected Post mapPost(Long postId) {
+        return postService.findById(postId);
     }
 
 }
